Review fixes
1) Parsing changes.
a) Process is divided into smaller parts that are easier to understand.
b) Makefile parsing process is separated from the rest of the logic and can be easily substituted.

2) Building changes.
a) Call stack was substituted with usual stack to avoid stack overflow exception.
b) Loop tracking data moved from the task itself to a wrapper, so we can run multiple builds simultaneously without the need to clean it up after.

3) Some other including removal of ConfirmedTask flag